# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BloodLink - Blood Bank Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new donor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: donorId
          in: query
          schema:
            type: integer
        - name: centerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, rejected, cancelled, completed]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"

    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Request an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment (confirm, reject, cancel)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available appointment slots
      parameters:
        - name: centerId
          in: query
          schema:
            type: integer
        - name: region
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlotsResponse"

  # DONATIONS
  /donations:
    get:
      operationId: listDonations
      tags: [donations]
      summary: List donation records
      parameters:
        - name: donorId
          in: query
          schema:
            type: integer
        - name: centerId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Donation records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationListResponse"

    post:
      operationId: recordDonation
      tags: [donations]
      summary: Record a new donation (barcode scan)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordDonationBody"
      responses:
        "201":
          description: Donation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"

  # BLOOD BAGS
  /blood-bags:
    get:
      operationId: listBloodBags
      tags: [blood-bags]
      summary: List blood bags in stock
      parameters:
        - name: bloodType
          in: query
          schema:
            type: string
        - name: centerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [available, reserved, transfused, expired, rejected]
        - name: expiringBefore
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Blood bags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodBagListResponse"
  /blood-bags/{id}:
    get:
      operationId: getBloodBag
      tags: [blood-bags]
      summary: Get blood bag details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blood bag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodBag"
  /blood-bags/scan/{barcode}:
    get:
      operationId: scanBloodBag
      tags: [blood-bags]
      summary: Scan blood bag by barcode
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blood bag data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodBag"
        "404":
          description: Not found
  /blood-bags/{id}/receive:
    post:
      operationId: receiveBloodBag
      tags: [blood-bags]
      summary: Receive a blood bag (blood bank inspection)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveBloodBagBody"
      responses:
        "200":
          description: Reception recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodBag"

  # STOCK
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List shortage alerts
      parameters:
        - name: bloodType
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved]
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertListResponse"

    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create shortage alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertBody"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update alert (resolve, acknowledge)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertBody"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"

  # INVOICES
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    RegisterBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        cin:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female]
        weight:
          type: number
        phone:
          type: string
        email:
          type: string
        password:
          type: string
        region:
          type: string
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        acceptedCharter:
          type: boolean
      required: [firstName, lastName, cin, dateOfBirth, gender, weight, phone, email, password, region, bloodType, acceptedCharter]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [donor, transfusion_center, blood_bank, hospital, clinic]
        bloodType:
          type: string
        region:
          type: string
        organizationName:
          type: string
      required: [id, firstName, lastName, email, role]

    # DONORS
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        donorId:
          type: integer
        donorName:
          type: string
        centerId:
          type: integer
        centerName:
          type: string
        region:
          type: string
        date:
          type: string
        time:
          type: string
        status:
          type: string
          enum: [pending, confirmed, rejected, cancelled, completed]
        notes:
          type: string
        createdAt:
          type: string
      required: [id, donorId, centerId, date, time, status]
    AppointmentListResponse:
      type: object
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
      required: [appointments, total]
    CreateAppointmentBody:
      type: object
      properties:
        centerId:
          type: integer
        date:
          type: string
        time:
          type: string
        notes:
          type: string
      required: [centerId, date, time]
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, rejected, cancelled, completed]
        assignedSlot:
          type: string
        notes:
          type: string
      required: [status]
    AvailableSlotsResponse:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required: [slots]
    TimeSlot:
      type: object
      properties:
        date:
          type: string
        time:
          type: string
        centerId:
          type: integer
        centerName:
          type: string
        available:
          type: boolean
      required: [date, time, centerId, available]

    # DONATIONS
    Donation:
      type: object
      properties:
        id:
          type: integer
        donorId:
          type: integer
        donorName:
          type: string
        centerId:
          type: integer
        centerName:
          type: string
        bloodBagId:
          type: integer
        bloodBagBarcode:
          type: string
        bloodType:
          type: string
        donationDate:
          type: string
        status:
          type: string
          enum: [collected, qualified, rejected]
      required: [id, donorId, centerId, bloodType, donationDate, status]
    DonationListResponse:
      type: object
      properties:
        donations:
          type: array
          items:
            $ref: "#/components/schemas/Donation"
        total:
          type: integer
      required: [donations, total]
    RecordDonationBody:
      type: object
      properties:
        donorId:
          type: integer
        barcode:
          type: string
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        collectionDate:
          type: string
        expirationDate:
          type: string
      required: [donorId, barcode, bloodType, collectionDate, expirationDate]

    # BLOOD BAGS
    BloodBag:
      type: object
      properties:
        id:
          type: integer
        barcode:
          type: string
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        donorId:
          type: integer
        centerId:
          type: integer
        centerName:
          type: string
        collectionDate:
          type: string
        expirationDate:
          type: string
        status:
          type: string
          enum: [available, reserved, transfused, expired, rejected]
        rejectionReason:
          type: string
        receivedAt:
          type: string
      required: [id, barcode, bloodType, collectionDate, expirationDate, status]
    BloodBagListResponse:
      type: object
      properties:
        bloodBags:
          type: array
          items:
            $ref: "#/components/schemas/BloodBag"
        total:
          type: integer
      required: [bloodBags, total]
    ReceiveBloodBagBody:
      type: object
      properties:
        temperatureOk:
          type: boolean
        integrityOk:
          type: boolean
        accepted:
          type: boolean
        rejectionReason:
          type: string
      required: [temperatureOk, integrityOk, accepted]

    # STOCK
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        centerId:
          type: integer
        centerName:
          type: string
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        region:
          type: string
        urgency:
          type: string
          enum: [normal, urgent, critical]
        message:
          type: string
        status:
          type: string
          enum: [active, resolved]
        acknowledgedBy:
          type: array
          items:
            type: integer
        createdAt:
          type: string
        resolvedAt:
          type: string
      required: [id, centerId, bloodType, region, urgency, message, status, createdAt]
    AlertListResponse:
      type: object
      properties:
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        total:
          type: integer
      required: [alerts, total]
    CreateAlertBody:
      type: object
      properties:
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        region:
          type: string
        urgency:
          type: string
          enum: [normal, urgent, critical]
        message:
          type: string
      required: [bloodType, region, urgency, message]
    UpdateAlertBody:
      type: object
      properties:
        status:
          type: string
          enum: [resolved]
        acknowledge:
          type: boolean
      required: [status]

    # INVOICES
...